fix 第一次安装拍照异常 fix bugly一些异常修复

+++ module-mine/src/main/java/com/rm/module_mine/activity/MimeFeedbackActivity.kt
@@ -176,7 +176,9 @@ class MimeFeedbackActivity : BaseVMActivity<MineActivityFeedbackBinding, MineFee
                 } else {
                     val cameraImagePath = mViewModel.photoHelp?.getCameraImagePath()
                     cameraImagePath?.let {
-                        compress(it, true)
+                        if (it.isNotEmpty()) {
+                            compress(it, true)
+                        }
                     }
                 }
             } else {
@@ -190,7 +192,9 @@ class MimeFeedbackActivity : BaseVMActivity<MineActivityFeedbackBinding, MineFee
                 } else {
                     val cameraImagePath = mViewModel.photoHelp?.getCameraImagePath()
                     cameraImagePath?.let {
-                        FileUtils.delete(it)
+                        if (it.isNotEmpty()) {
+                            FileUtils.delete(it)
+                        }
                     }
                 }
             }
